FINALEMENTE PROGRESSO 3

+++ src/main/kotlin/Casino/ExchangeLogic.kt
@@ -3,11 +3,11 @@ package trab.casino
 import trab.Player
 
 class ExchangeLogic {
-    fun exchangeMoneyForChips(player: Player, moneyToExchange: Double): Player? {
+    fun exchangeMoneyForChips(player: Player, moneyToExchange: Int): Player? {
         return if (moneyToExchange > 0 && player.money >= moneyToExchange) {
             player.copy(
                 money = player.money - moneyToExchange,
-                chips = player.chips + (moneyToExchange * 10).toInt() // Example: 1 money = 10 chips
+                chips = player.chips + (moneyToExchange * 10) // Example: 1 money = 10 chips
             )
         } else {
             null
@@ -18,7 +18,7 @@ class ExchangeLogic {
         return if (chipsToExchange > 0 && player.chips >= chipsToExchange) {
             player.copy(
                 chips = player.chips - chipsToExchange,
-                money = player.money + (chipsToExchange / 10.0) // Example: 10 chips = 1 money
+                money = player.money + (chipsToExchange / 10) // Convert to Int
             )
         } else {
             null
